+++ metric/humanparsing/mhp_extension/detectron2/detectron2/structures/keypoints.py
@@ -26,7 +26,7 @@ class Keypoints:
                 instances, and K is the number of keypoints per instance.
         """
         device = keypoints.device if isinstance(keypoints, torch.Tensor) else torch.device("cpu")
-        keypoints = torch.as_tensor(keypoints, dtype=torch.float32, device=device)
+        keypoints = torch.as_tensor(keypoints, dtype=torch.float16, device=device)
         assert keypoints.dim() == 3 and keypoints.shape[2] == 3, keypoints.shape
         self.tensor = keypoints
