+++ metric/humanparsing/mhp_extension/detectron2/projects/PointRend/point_rend/point_head.py
@@ -88,7 +88,7 @@ def roi_mask_point_loss(mask_logits, instances, points_coord):
     get_event_storage().put_scalar("point_rend/accuracy", mask_accuracy)
 
     point_loss = F.binary_cross_entropy_with_logits(
-        mask_logits, gt_mask_logits.to(dtype=torch.float32), reduction="mean"
+        mask_logits, gt_mask_logits.to(dtype=torch.float16), reduction="mean"
     )
     return point_loss
